replace nulls with Optional
1. return Optional<> instead of nulls
2. throw the IOException to ClientRunner

+++ client/src/main/java/sg/edu/ntu/nfs/client/CacheHandler.java
@@ -1,5 +1,8 @@
 package sg.edu.ntu.nfs.client;
 
+import java.io.IOException;
+import java.util.Optional;
+
 public class CacheHandler {
 
     static Cache cache = new Cache(); // only one instance
@@ -17,54 +20,59 @@ public class CacheHandler {
      * @param file_path file path on server
      * @return file content in bytes
      */
-    public byte[] getFile(String file_path) throws NullPointerException {
+    public Optional<byte[]> getFile(String file_path) throws IOException {
         CacheEntry entry = cache.getFile(file_path);
+        Optional<byte[]> opt_content;
 
         // not in cache
         if (entry == null) {
-            byte[] file_content = stub.requestFile(file_path);
-
+            opt_content = stub.requestFile(file_path);
             // file found on server
-            if (file_content != null) {
+            if (opt_content.isPresent()) {
+                byte[] file_content = opt_content.get();
                 long t_mclient = System.currentTimeMillis();
                 long t_c = System.currentTimeMillis();
                 cache.addFile(file_path, file_content, t_mclient, t_c);
-                return file_content;
-            } else
-                return null;
+            }
         }
-
-        // in cache
         else {
             // check freshness upon access
             if (System.currentTimeMillis() - entry.getTc() >= cache.getFreshT()) {
-                long[] attr = stub.getAttr((file_path));
-                long t_mserver = attr[0];
+                Optional<long[]> opt_attr = stub.getAttr((file_path));
 
-                // invalid entry
-                if (entry.getTmclient() < t_mserver) {
-                    // update entry
-                    byte[] file_content = stub.requestFile(file_path);
+                if (opt_attr.isPresent()) {
+                    long[] attr = opt_attr.get();
+                    long t_mserver = attr[0];
 
-                    // file still on server
-                    if (file_content != null) {
-                        long t_mclient = System.currentTimeMillis();
-                        long t_c = System.currentTimeMillis();
-                        cache.replaceFile(file_path, file_content, t_mclient, t_c);
+                    // invalid entry
+                    if (entry.getTmclient() < t_mserver) {
+                        // update entry
+                        Optional<byte[]> opt_file = stub.requestFile(file_path);
+
+                        if (opt_file.isPresent()) {
+                            byte[] file_content = opt_file.get();
+
+                            // file still on server
+                            if (file_content != null) {
+                                long t_mclient = System.currentTimeMillis();
+                                long t_c = System.currentTimeMillis();
+                                cache.replaceFile(file_path, file_content, t_mclient, t_c);
+                            }
+                            // file no longer available on server
+                            else
+                                cache.removeFile(file_path);
+                        }
                     }
-                    // file no longer available on server
-                    else
-                        cache.removeFile(file_path);
                 }
             }
+            opt_content = Optional.of(entry.getFileContent());
         }
-        return entry.getFileContent();
+        return opt_content;
     }
 
     public Cache getCache(){
         return cache;
     }
-
 
 
     /**
